[FIX] 도장판 상세 헤더 부분 레이아웃 수정
issue: #55

* 도장판 타이틀 우측의 날짜 chip에 상단 padding 값 추가
